make sure we clear queue once conversation changes
